Add perceptron.py, chagen file --> features.txt
Perceptron.py in its current will parse through the given data files and
generate arrays of boolean features as suggested by Liang. The Bias
feature is always set to 1. This should be explored to improve
performance later. A feature mapping is generated so that phrases such
as "married-civ..." can be mapped to an integer index in the boolean
numpy array. This is accomplished through featDict() function, which
also can take a custome features file. This should not be done in most
cases.

File to feature will take a filename and the generated feature
dictionary and will populate a 2d array with each entry's respective
feature vector. there's a bit of hacky stuff going on. Uses a default
python list solely because .append is a method and works better than
rebuilding a numpy array every time. This function also optionally
returns a list of the value of someone's income. If >50k then the list
value is 1, if <=50k then value is -1. Order in the feature vector and
in the truth list is maintained.